isp-customer-service/chatbot_core/requirements.txt
streamlit>=1.30.0
litellm>=1.80.5
openai>=1.0.0
langchain>=0.1.0
langgraph>=0.0.20
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pyyaml>=6.0
pydantic>=2.0.0
colorama>=0.4.6
python-dotenv>=1.0.0